agentic_ui_tester/requirements.txt
playwright
langchain
langchain-ollama
python-dotenv